+++ src/components/index.ts
@@ -7,5 +7,5 @@ export { default as InputText } from './InputText.vue'
 export { default as InputFile } from './InputFile.vue'
 export { default as InputTextArea } from './InputTextArea.vue'
 export { default as Modal } from './Modal.vue'
-export { default as WishlistHeader } from './WishlistHeader.vue'
+export { default as FormWishlist } from './FormWishlist.vue'
 export { default as WishlistItem } from './WishlistItem.vue'
